+++ src/core/model/transform.hpp
@@ -28,7 +28,7 @@ public:
 
     virtual QString type_name_human() const override { return tr("Transform"); }
 
-    QTransform transform_matrix(FrameTime f) const;
+    QTransform transform_matrix(FrameTime f, bool auto_orient = false) const;
     void set_transform_matrix(const QTransform& t);
     void copy(Transform* other);
 };